+++ src/zodbpickle/tests/test_pickle_2.py
@@ -1,13 +1,49 @@
+import cStringIO
+import io
 import unittest
 from cStringIO import StringIO
 
-from .pickletester_2 import _AbstractPickleTests
-from .pickletester_2 import _AbstractPickleModuleTests
-from .pickletester_2 import _AbstractPersistentPicklerTests
-from .pickletester_2 import _AbstractPicklerUnpicklerObjectTests
+from .pickletester_2 import (AbstractPickleTests,
+                             AbstractPickleModuleTests,
+                             AbstractPersistentPicklerTests,
+                             AbstractPicklerUnpicklerObjectTests,
+                             BigmemPickleTests)
+
+from test import test_support
+
+class cStringIOMixin:
+    output = input = cStringIO.StringIO
+
+    def close(self, f):
+        pass
+
+class BytesIOMixin:
+    output = input = io.BytesIO
+
+    def close(self, f):
+        pass
+
+class FileIOMixin:
+
+    def output(self):
+        return open(test_support.TESTFN, 'wb+')
+
+    def input(self, data):
+        f = open(test_support.TESTFN, 'wb+')
+        try:
+            f.write(data)
+            f.seek(0)
+            return f
+        except:
+            f.close()
+            raise
+
+    def close(self, f):
+        f.close()
+        test_support.unlink(test_support.TESTFN)
 
 
-class PickleTests(_AbstractPickleTests, _AbstractPickleModuleTests):
+class PickleTests(AbstractPickleTests, AbstractPickleModuleTests):
 
     def dumps(self, arg, proto=0, fast=0):
         from zodbpickle.pickle_2 import dumps
@@ -27,13 +63,13 @@ class PickleTests(_AbstractPickleTests, _AbstractPickleModuleTests):
     error = KeyError
 
 
-class PicklerTests(_AbstractPickleTests):
+class PicklerTests(AbstractPickleTests):
 
     error = KeyError
 
     def dumps(self, arg, proto=0, fast=0):
         from zodbpickle.pickle_2 import Pickler
-        f = StringIO()
+        f = cStringIO.StringIO()
         p = Pickler(f, proto)
         if fast:
             p.fast = fast
@@ -43,19 +79,19 @@ class PicklerTests(_AbstractPickleTests):
 
     def loads(self, buf):
         from zodbpickle.pickle_2 import Unpickler
-        f = StringIO(buf)
+        f = cStringIO.StringIO(buf)
         u = Unpickler(f)
         return u.load()
 
 
-class PersPicklerTests(_AbstractPersistentPicklerTests):
+class PersPicklerTests(AbstractPersistentPicklerTests):
 
     def dumps(self, arg, proto=0, fast=0):
         from zodbpickle.pickle_2 import Pickler
         class PersPickler(Pickler):
             def persistent_id(subself, obj):
                 return self.persistent_id(obj)
-        f = StringIO()
+        f = cStringIO.StringIO()
         p = PersPickler(f, proto)
         if fast:
             p.fast = fast
@@ -68,22 +104,35 @@ class PersPicklerTests(_AbstractPersistentPicklerTests):
         class PersUnpickler(Unpickler):
             def persistent_load(subself, obj):
                 return self.persistent_load(obj)
-        f = StringIO(buf)
+        f = cStringIO.StringIO(buf)
         u = PersUnpickler(f)
         return u.load()
 
 
-class PicklerUnpicklerObjectTests(_AbstractPicklerUnpicklerObjectTests):
+class PicklerUnpicklerObjectTests(AbstractPicklerUnpicklerObjectTests):
 
     @property
     def pickler_class(self):
-        from zodbpickle._pickle import Pickler
+        from zodbpickle.pickle_2 import Pickler
         return  Pickler
 
     @property
     def unpickler_class(self):
-        from zodbpickle._pickle import Unpickler
+        from zodbpickle.pickle_2 import Unpickler
         return  Unpickler
+
+
+class PickleBigmemPickleTests(BigmemPickleTests):
+
+    def dumps(self, arg, proto=0, fast=0):
+        from zodbpickle import pickle_2
+        # Ignore fast
+        return pickle_2.dumps(arg, proto)
+
+    def loads(self, buf):
+        from zodbpickle import pickle_2
+        # Ignore fast
+        return pickle_2.loads(buf)
 
 
 class cPickleBase(object):
@@ -94,8 +143,8 @@ class cPickleBase(object):
         return BadPickleGet
 
 
-class cPickleTests(_AbstractPickleTests,
-                   _AbstractPickleModuleTests,
+class cPickleTests(AbstractPickleTests,
+                   AbstractPickleModuleTests,
                    cPickleBase,
                   ):
     def setUp(self):
@@ -110,78 +159,111 @@ class cPickleTests(_AbstractPickleTests,
         return _pickle
 
 
-class cPicklePicklerTests(_AbstractPickleTests, cPickleBase):
+class cPicklePicklerTests(AbstractPickleTests, cPickleBase):
 
     def dumps(self, arg, proto=0):
-        from zodbpickle._pickle import Pickler
-        f = StringIO()
-        p = Pickler(f, proto)
-        p.dump(arg)
-        f.seek(0)
-        return f.read()
+        from zodbpickle import _pickle
+        f = self.output()
+        try:
+            p = _pickle.Pickler(f, proto)
+            p.dump(arg)
+            f.seek(0)
+            return f.read()
+        finally:
+            self.close(f)
 
     def loads(self, buf):
-        from zodbpickle._pickle import Unpickler
-        f = StringIO(buf)
-        p = Unpickler(f)
-        return p.load()
+        from zodbpickle import _pickle
+        f = self.input(buf)
+        try:
+            p = _pickle.Unpickler(f)
+            return p.load()
+        finally:
+            self.close(f)
+
+class cStringIOCPicklerTests(cStringIOMixin, cPicklePicklerTests):
+    pass
+
+class BytesIOCPicklerTests(BytesIOMixin, cPicklePicklerTests):
+    pass
+
+class FileIOCPicklerTests(FileIOMixin, cPicklePicklerTests):
+    pass
 
 
-class cPickleListPicklerTests(_AbstractPickleTests, cPickleBase):
+class cPickleListPicklerTests(AbstractPickleTests, cPickleBase):
 
     def dumps(self, arg, proto=0):
-        from zodbpickle._pickle import Pickler
-        p = Pickler(proto)
+        from zodbpickle import _pickle
+        p = _pickle.Pickler(proto)
         p.dump(arg)
         return p.getvalue()
 
     def loads(self, *args):
-        from zodbpickle._pickle import Unpickler
-        f = StringIO(args[0])
-        p = Unpickler(f)
-        return p.load()
+        from zodbpickle import _pickle
+        f = self.input(args[0])
+        try:
+            p = _pickle.Unpickler(f)
+            return p.load()
+        finally:
+            self.close(f)
+
+class cStringIOCPicklerListTests(cStringIOMixin, cPickleListPicklerTests):
+    pass
+
+class BytesIOCPicklerListTests(BytesIOMixin, cPickleListPicklerTests):
+    pass
+
+class FileIOCPicklerListTests(FileIOMixin, cPickleListPicklerTests):
+    pass
 
 
-class cPickleFastPicklerTests(_AbstractPickleTests, cPickleBase):
+class cPickleFastPicklerTests(AbstractPickleTests, cPickleBase):
 
     def dumps(self, arg, proto=0):
-        from zodbpickle._pickle import Pickler
-        f = StringIO()
-        p = Pickler(f, proto)
-        p.fast = 1
-        p.dump(arg)
-        f.seek(0)
-        return f.read()
+        from zodbpickle import _pickle
+        f = self.output()
+        try:
+            p = _pickle.Pickler(f, proto)
+            p.fast = 1
+            p.dump(arg)
+            f.seek(0)
+            return f.read()
+        finally:
+            self.close(f)
 
     def loads(self, *args):
-        from zodbpickle._pickle import Unpickler
-        f = StringIO(args[0])
-        p = Unpickler(f)
-        return p.load()
+        from zodbpickle import _pickle
+        f = self.input(args[0])
+        try:
+            p = _pickle.Unpickler(f)
+            return p.load()
+        finally:
+            self.close(f)
 
     def test_recursive_list(self):
         self.assertRaises(ValueError,
-                          _AbstractPickleTests.test_recursive_list,
+                          AbstractPickleTests.test_recursive_list,
                           self)
 
     def test_recursive_tuple(self):
         self.assertRaises(ValueError,
-                          _AbstractPickleTests.test_recursive_tuple,
+                          AbstractPickleTests.test_recursive_tuple,
                           self)
 
     def test_recursive_inst(self):
         self.assertRaises(ValueError,
-                          _AbstractPickleTests.test_recursive_inst,
+                          AbstractPickleTests.test_recursive_inst,
                           self)
 
     def test_recursive_dict(self):
         self.assertRaises(ValueError,
-                          _AbstractPickleTests.test_recursive_dict,
+                          AbstractPickleTests.test_recursive_dict,
                           self)
 
     def test_recursive_multi(self):
         self.assertRaises(ValueError,
-                          _AbstractPickleTests.test_recursive_multi,
+                          AbstractPickleTests.test_recursive_multi,
                           self)
 
     def test_nonrecursive_deep(self):
@@ -195,7 +277,17 @@ class cPickleFastPicklerTests(_AbstractPickleTests, cPickleBase):
         b = self.loads(self.dumps(a))
         self.assertEqual(a, b)
 
-class cPicklePicklerUnpicklerObjectTests(_AbstractPicklerUnpicklerObjectTests):
+class cStringIOCPicklerFastTests(cStringIOMixin, cPickleFastPicklerTests):
+    pass
+
+class BytesIOCPicklerFastTests(BytesIOMixin, cPickleFastPicklerTests):
+    pass
+
+class FileIOCPicklerFastTests(FileIOMixin, cPickleFastPicklerTests):
+    pass
+
+
+class cPicklePicklerUnpicklerObjectTests(AbstractPicklerUnpicklerObjectTests):
 
     @property
     def pickler_class(self):
@@ -207,10 +299,21 @@ class cPicklePicklerUnpicklerObjectTests(_AbstractPicklerUnpicklerObjectTests):
         from zodbpickle._pickle import Unpickler
         return Unpickler
 
+class cPickleBigmemPickleTests(BigmemPickleTests):
+
+    def dumps(self, arg, proto=0, fast=0):
+        from zodbpickle import _pickle
+        # Ignore fast
+        return _pickle.dumps(arg, proto)
+
+    def loads(self, buf):
+        from zodbpickle import _pickle
+        # Ignore fast
+        return _pickle.loads(buf)
+
 
 class Node(object):
     pass
-
 
 class cPickleDeepRecursive(unittest.TestCase):
 
@@ -218,21 +321,21 @@ class cPickleDeepRecursive(unittest.TestCase):
         # This should raise a RecursionLimit but in some
         # platforms (FreeBSD, win32) sometimes raises KeyError instead,
         # or just silently terminates the interpreter (=crashes).
-        from zodbpickle._pickle import dumps
+        from zodbpickle import _pickle
         nodes = [Node() for i in range(500)]
         for n in nodes:
             n.connections = list(nodes)
             n.connections.remove(n)
-        self.assertRaises((AttributeError, RuntimeError), dumps, n)
+        self.assertRaises((AttributeError, RuntimeError), _pickle.dumps, n)
 
     def test_issue3179(self):
         # Safe test, because I broke this case when fixing the
         # behaviour for the previous test.
-        from zodbpickle._pickle import dumps
+        from zodbpickle import _pickle
         res=[]
         for x in range(1,2000):
             res.append(dict(doc=x, similar=[]))
-        dumps(res)
+        _pickle.dumps(res)
 
 
 def test_suite():
@@ -242,10 +345,22 @@ def test_suite():
         unittest.makeSuite(PicklerTests),
         unittest.makeSuite(PersPicklerTests),
         unittest.makeSuite(PicklerUnpicklerObjectTests),
+        unittest.makeSuite(PickleBigmemPickleTests),
+        
         unittest.makeSuite(cPickleTests),
-        unittest.makeSuite(cPicklePicklerTests),
-        unittest.makeSuite(cPickleListPicklerTests),
-        unittest.makeSuite(cPickleFastPicklerTests),
+        unittest.makeSuite(cStringIOCPicklerTests),
+        unittest.makeSuite(BytesIOCPicklerTests),
+        unittest.makeSuite(FileIOCPicklerTests),        
+        unittest.makeSuite(cStringIOCPicklerListTests),
+        unittest.makeSuite(BytesIOCPicklerListTests),
+        unittest.makeSuite(FileIOCPicklerListTests),
+        unittest.makeSuite(cStringIOCPicklerFastTests),
+        unittest.makeSuite(BytesIOCPicklerFastTests),
+        unittest.makeSuite(FileIOCPicklerFastTests),
         unittest.makeSuite(cPickleDeepRecursive),
         unittest.makeSuite(cPicklePicklerUnpicklerObjectTests),
+        unittest.makeSuite(cPickleBigmemPickleTests),
     ))
+
+if __name__ == '__main__':
+    test_support.run_unittest(test_suite())